requirements.txt
<<<<<<< HEAD
Flask
Flask-CORS
gunicorn
google-generativeai
python-pptx
requests
werkzeug
python-dotenv
pdfminer.six
=======
Flask
Flask-CORS
gunicorn
google-generativeai
python-pptx
requests
werkzeug
python-dotenv
pdfminer.six
>>>>>>> b94a9ca662e4d003bf206be789b87968fc9c6eb5
python-docx